Do not wait for settrace() to finish until wait_for_attach(). (#628)
(helps #545)

We move the settrace() into a thread, but make sure to wait for the thread to finish in wait_for_attach().

My only concern with the solution is that the current thread will not get debugged if enable_attach() is called but wait_for_attach() is not (and no attach happens right away). I'm going to address this in a follow-up PR by monkey-patching sys.settrace temporarily.