Fix #1612: Parse and expose DAP capabilities
Add ptvsd.adapter.contract module as a single access point for capabilities and other immutable IDE and server metadata.

Change the semantics of Channels.server() to enable checks for disconnected server.

Fix "disconnect" and "terminate" handling when not connected to debug server.

Do best-effort handling of "ptvsd_systemInfo" if not connected to debug server.